+++ CHANGELOG.md
@@ -8,15 +8,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- Added x-reference-type property to define the type of the reference
-- Added exclusions of validation errors on application level
-- Added support for links to java implementations using linkSrcDir property
 
 ### Changed
 
 ### Deprecated
 
 ### Removed
+
+### Fixed
+
+## [1.12.0] - 2025-04-13
+
+### Added
+- Exclusions of validation errors
+- x-reference-type property to define the type of the reference
+- Support for links to java implementations using linkSrcDir property
+
+### Removed
+- test-real as it is not needed any more
 
 ### Fixed